v11.2: harden accumulator int casts to long (overflow safety)
Belt-and-suspenders defense after the v11.1 cap-always-on cleanup. The
v11.1 commit removed the path that triggered int32 overflow in practice
(unbounded prey growth produced wholeBirths > 2.1B around day 50),
but the cast itself was still int. If a future code path ever
produces float counts above int.MaxValue — through misconfiguration,
bug, or intentional huge-population scenarios — the cast would still
wrap to a negative value and corrupt the simulation.

Changes EcosystemSimulator.cs in four places — all the
"int wholeFoo = (int)Math.Floor(accumulated)" patterns:

1. ProcessFeedingWithAccumulator (predation accumulator, line ~764)
2. ApplyConditionDeath (condition death accumulator, line ~941)
3. ApplyReproduction (birth accumulator, line ~1098)
4. ApplyNaturalDeathWithAccumulator (natural death accumulator, line ~1156)

Each becomes "long wholeFoo = (long)Math.Floor(accumulated)" with the
matching "Math.Min(wholeFoo, (long)sp.Population)" cap also widened.

Downstream Last* trackers stay float (implicit long → float widening
is a no-op for normal magnitudes), and StepRecord BirthsT1/T2 are
already long. No semantic change at normal magnitudes.

float precision is still the underlying limit (~24 bits of integer
precision past 16M), but float can represent up to ~3.4e38, so the
practical safety margin is enormous compared to int32's 2.1B.

+++ Assets/scripts/Simulation/EcosystemSimulator.cs
@@ -760,12 +760,12 @@ public class EcosystemSimulator
                 _predationAccumulators[p.FullName] += preyLost;
                 float accumulated = _predationAccumulators[p.FullName];
 
-                // Extract whole deaths
-                int wholeDeaths = (int)Math.Floor(accumulated);
+                // Extract whole deaths (long to avoid int32 overflow at large prey pops)
+                long wholeDeaths = (long)Math.Floor(accumulated);
                 _predationAccumulators[p.FullName] = accumulated - wholeDeaths;
 
                 // Apply deaths
-                wholeDeaths = Math.Min(wholeDeaths, (int)p.Population);
+                wholeDeaths = Math.Min(wholeDeaths, (long)p.Population);
                 float oldPop = p.Population;
                 p.Population = Math.Max(0f, p.Population - wholeDeaths);
                 LastEatenT1 += wholeDeaths;
@@ -935,12 +935,13 @@ public class EcosystemSimulator
         float severity = (sp.DeathThreshold - sp.Condition) / sp.DeathThreshold;
         float rawDeaths = sp.Population * severity * sp.DeathRate * BiologyStep;
 
-        // Accumulator pattern — fractional deaths carry over between days
+        // Accumulator pattern — fractional deaths carry over between days.
+        // long (not int) to avoid overflow at extreme populations.
         _conditionDeathAccumulators[sp.FullName] += rawDeaths;
         float accumulated = _conditionDeathAccumulators[sp.FullName];
-        int wholeDeaths = (int)Math.Floor(accumulated);
+        long wholeDeaths = (long)Math.Floor(accumulated);
         _conditionDeathAccumulators[sp.FullName] = accumulated - wholeDeaths;
-        wholeDeaths = Math.Min(wholeDeaths, (int)sp.Population);
+        wholeDeaths = Math.Min(wholeDeaths, (long)sp.Population);
 
         if (wholeDeaths > 0)
         {
@@ -1095,7 +1096,10 @@ public class EcosystemSimulator
         _birthAccumulators[sp.FullName] += births;
         float accumulated = _birthAccumulators[sp.FullName];
 
-        int wholeBirths = (int)Math.Floor(accumulated);
+        // long (not int) to avoid int32 overflow if a misconfigured scenario
+        // produces birth counts > 2.1B in a single step. With cap-always-on (v11.1)
+        // this should not happen in practice, but the safety guard is cheap.
+        long wholeBirths = (long)Math.Floor(accumulated);
         _birthAccumulators[sp.FullName] = accumulated - wholeBirths;
 
         float oldPop = sp.Population;
@@ -1153,11 +1157,12 @@ public class EcosystemSimulator
         _naturalDeathAccumulators[sp.FullName] += deaths;
         float accumulated = _naturalDeathAccumulators[sp.FullName];
 
-        int wholeDeaths = (int)Math.Floor(accumulated);
+        // long (not int) for overflow safety at extreme populations.
+        long wholeDeaths = (long)Math.Floor(accumulated);
         _naturalDeathAccumulators[sp.FullName] = accumulated - wholeDeaths;
 
         // Cap deaths at population
-        wholeDeaths = Math.Min(wholeDeaths, (int)sp.Population);
+        wholeDeaths = Math.Min(wholeDeaths, (long)sp.Population);
 
         if (wholeDeaths > 0)
         {
